+++ js/setup.js
@@ -1,0 +1,38 @@
+const INDEX_COUNT = 25;
+
+const MAX_COMMENTS = 200;
+
+const likeCount = {
+  min: 15,
+  max: 200,
+};
+
+const photoDescriptions = [
+  'Расслабься и смотри',
+  'Кусочек самого лучшего путешествия',
+  'Неплохая локация для съемок',
+  'Просто оставлю это сдесь',
+  'Иногда хочется просто хочется такой же обстановки',
+  'Такие места вдохновляют',
+  'Были тут?',
+  'Об этом мечтает каждый',
+  'Чего бы это не стоило',
+  'У меня есть в запасе последний глоток тишины',
+  'Всегда новые локации!',
+  'Самый лучший день',
+  'Путешествовать и жить гораздо интереснее, если следуешь внезапно возникающим импульсам',
+  'На заставку',
+  'Тут бывает и не такое',
+  'Малая часть того, чем бы я хотел поделиться',
+  'Это самое потрясающее место!',
+  'Это не может быть правдой!',
+  'Такого в наших краях не увидишь',
+  'Без фильтра',
+  'Ещё одни тёплые выходные',
+  'Вся наша жизнь подчиняется природным ритмам',
+  'Есть в жизни счастье',
+  'Вперед исследовать!',
+  'Мое любимое место',
+];
+
+export {INDEX_COUNT, MAX_COMMENTS, likeCount, photoDescriptions};
